third_party/WebKit/LayoutTests/fast/speech/bubble-position-expected.txt

CONSOLE WARNING: The 'x-webkit-speech' input field attribute is deprecated. Please use the JavaScript API instead.
Tests for speech button click with <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Clicking the speech button.
PASS rect is inside the iframe
PASS successfullyParsed is true

TEST COMPLETE
